Moves current functions from web_scraper.py to get_images.py
web_scraper.py contained two versions of fetch_images(),
a previously used one that was commented out and
a current one along with other currently used functions, these
were moved into get_images.py. web_scraper.py doesn't work as is,
other dependencies need to be imported but it will serve
as an optional function call in case the strikeforce api
no longer works in the future.
